requirements.txt
ipykernel
pandas
numpy
tqdm
jupyter
matplotlib
seaborn
scikit-learn
gdown
colorlog
nltk
wordcloud
pyarrow
mlflow<3
dagshub
python-dotenv
packaging<25
imblearn
imbalanced-learn
xgboost
lightgbm
optuna
Flask
Flask-Cors
gunicorn
dvc
python-box
ensure